requirements.txt
julep
gradio
python-dotenv 